Fix noise radius implementation
Before it was not implemented correctly, as the noise was added but not used. Now the noise is added to the spot radius, and if the resulting radius is less than 1, it is set to 1 to avoid negative or zero radius. This results in spots of different sizes, as the noise sd grows.
Also, I removed the flag noise_radius, as it is redundant with noise_radius_sd = 0 meaning no noise.