Don't add metadata to sharp treeline landscapes
When we create sharp treeline landscapes as a basis for other landscape types, we don't need metadata (like treeline position) to be added to the output. This is because the sharp treeline landscape is just an intermediate step and the metadata will be added later when we create the final landscape type (like scatter or fingers). If metadata is added here, we get an error because the output is a list rather than a landscape matrix.